+++ CHANGELOG.md
@@ -10,7 +10,10 @@ Alternatively view it [online on GitHub](https://github.com/cumulusmx/CumulusMX/
 ---
 ---
 
-## [4.5.0 \[b4093\]][19] - 2025-05-15
+## [4.5.0 \[b4094\]][19] - 2025-05-16
+
+### b4094 changes
+- Ecowitt camera web tag warning messages changed to plain messages
 
 ### b4093 changes
 - Davis station: Fix extra T/H/DP/Soil Moist/Soil Temp/Leaf Wetness being wrongly set to null
@@ -830,4 +833,4 @@ Initial release of Cumulus MX which now runs under Microsoft .NET 8.0 and remove
 [16]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4086
 [17]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4087
 [18]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4088
-[19]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4093
+[19]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4094